refactor: dbus_method fixture to conftest
Make this available to other tests as well
