+++ tests/unit/test_core/test_calls.py
@@ -6,18 +6,6 @@ from wakepy.core import DBusAddress, DBusMethod, DBusMethodCall
 
 
 @pytest.fixture
-def service():
-    return DBusAddress(path="/foo", service="wakepy.foo", interface="/foo")
-
-
-@pytest.fixture
-def method(service):
-    return DBusMethod(
-        name="test-method", signature="isi", params=("first", "second", "third")
-    ).of(service)
-
-
-@pytest.fixture
 def method_without_params(service):
     service = DBusAddress(path="/foo", service="wakepy.foo", interface="/foo")
     return DBusMethod(
@@ -26,43 +14,43 @@ def method_without_params(service):
     ).of(service)
 
 
-def test_dbusmethod_args_none(method: DBusMethod):
-    call = DBusMethodCall(method, args=None)
+def test_dbusmethod_args_none(dbus_method: DBusMethod):
+    call = DBusMethodCall(dbus_method, args=None)
     assert call.args == tuple()
 
 
-def test_dbusmethod_args_missing(method: DBusMethod):
-    call = DBusMethodCall(method)
+def test_dbusmethod_args_missing(dbus_method: DBusMethod):
+    call = DBusMethodCall(dbus_method)
     assert call.args == tuple()
 
 
-def test_dbusmethod_args_tuple(method: DBusMethod):
+def test_dbusmethod_args_tuple(dbus_method: DBusMethod):
     args = (1, "2", 3)
-    call = DBusMethodCall(method, args=args)
+    call = DBusMethodCall(dbus_method, args=args)
     assert call.args == args
 
 
-def test_dbusmethod_args_tuple_too_long(method: DBusMethod):
+def test_dbusmethod_args_tuple_too_long(dbus_method: DBusMethod):
     args = (1, "2", 3, 4)
     with pytest.raises(
         ValueError,
         match=re.escape("Expected args to have 3 items! (has: 4)"),
     ):
-        DBusMethodCall(method, args=args)
+        DBusMethodCall(dbus_method, args=args)
 
 
-def test_dbusmethod_args_tuple_too_short(method: DBusMethod):
+def test_dbusmethod_args_tuple_too_short(dbus_method: DBusMethod):
     args = (1, "2")
     with pytest.raises(
         ValueError,
         match=re.escape("Expected args to have 3 items! (has: 2)"),
     ):
-        DBusMethodCall(method, args=args)
+        DBusMethodCall(dbus_method, args=args)
 
 
-def test_dbusmethod_args_list(method: DBusMethod):
+def test_dbusmethod_args_list(dbus_method: DBusMethod):
     args = [1, "2", 3]
-    call = DBusMethodCall(method, args=args)
+    call = DBusMethodCall(dbus_method, args=args)
     assert call.args == (1, "2", 3)
 
 
@@ -81,27 +69,27 @@ def test_dbusmethod_args_dict_method_without_params(
         DBusMethodCall(method_without_params, args=args)
 
 
-def test_dbusmethod_args_dict_too_many_keys(method: DBusMethod):
+def test_dbusmethod_args_dict_too_many_keys(dbus_method: DBusMethod):
     args = dict(first=1, second="2", third=3, fourth=4)
 
     with pytest.raises(
         ValueError,
         match=re.escape("Expected args to have 3 items! (has: 4)"),
     ):
-        DBusMethodCall(method, args=args)
+        DBusMethodCall(dbus_method, args=args)
 
 
-def test_dbusmethod_args_dict_too_few_keys(method: DBusMethod):
+def test_dbusmethod_args_dict_too_few_keys(dbus_method: DBusMethod):
     args = dict(first=1, second="2")
 
     with pytest.raises(
         ValueError,
         match=re.escape("Expected args to have 3 items! (has: 2)"),
     ):
-        DBusMethodCall(method, args=args)
+        DBusMethodCall(dbus_method, args=args)
 
 
-def test_dbusmethod_args_dict_wrong_keys(method: DBusMethod):
+def test_dbusmethod_args_dict_wrong_keys(dbus_method: DBusMethod):
     args = dict(first=1, second="2", fifth="2")
 
     with pytest.raises(
@@ -111,24 +99,24 @@ def test_dbusmethod_args_dict_wrong_keys(method: DBusMethod):
             "Expected: ('first', 'second', 'third'). Got: ('first', 'second', 'fifth')"
         ),
     ):
-        DBusMethodCall(method, args=args)
+        DBusMethodCall(dbus_method, args=args)
 
 
-def test_dbusmethod_args_dict(method: DBusMethod):
+def test_dbusmethod_args_dict(dbus_method: DBusMethod):
     args = dict(first=1, second="2", third=3)
-    call = DBusMethodCall(method, args=args)
+    call = DBusMethodCall(dbus_method, args=args)
     assert call.args == (1, "2", 3)
 
 
-def test_dbusmethod_args_dict_different_order(method: DBusMethod):
+def test_dbusmethod_args_dict_different_order(dbus_method: DBusMethod):
     args = dict(third=3, first=1, second="2")
-    call = DBusMethodCall(method, args=args)
+    call = DBusMethodCall(dbus_method, args=args)
     assert call.args == (1, "2", 3)
 
 
-def test_dbusmethod_get_kwargs(method: DBusMethod):
+def test_dbusmethod_get_kwargs(dbus_method: DBusMethod):
     args = (1, "2", 3)
-    call = DBusMethodCall(method, args=args)
+    call = DBusMethodCall(dbus_method, args=args)
     assert call.get_kwargs() == dict(first=1, second="2", third=3)
 
 
@@ -139,7 +127,7 @@ def test_dbusmethod_get_kwargs_noparams(method_without_params: DBusMethod):
     assert call.get_kwargs() is None
 
 
-def test_dbusmethod_string_representation(method: DBusMethod):
+def test_dbusmethod_string_representation(dbus_method: DBusMethod):
     args = (1, "2", 3)
-    call = DBusMethodCall(method, args=args)
+    call = DBusMethodCall(dbus_method, args=args)
     assert call.__repr__() == "<wakepy.foo (1, '2', 3) | bus: SESSION>"
